feat(replicator): feature-gate destinations
